requirements.txt
click >= 8
pyyaml
pytest
